feature/reservation move DayOfWeek ENUM to cafe domain#14
