+++ src/main/java/org/example/stayd/domain/cafe/model/DayOfWeek.java
@@ -1,4 +1,4 @@
-package org.example.stayd.domain.reservation.model;
+package org.example.stayd.domain.cafe.model;
 
 public enum DayOfWeek {
     MON, TUE, WED, THU, FRI, SAT, SUN;
@@ -7,7 +7,7 @@ public enum DayOfWeek {
         return DayOfWeek.valueOf(stringValue);
     }
 
-    public String toValue() {
+    public String toStringValue() {
         return name();
     }
 }
