+++ src/main/java/org/example/stayd/domain/reservation/model/Reservation.java
@@ -5,15 +5,13 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
-import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Optional;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.stayd.common.YesNoBooleanConverter;
+import org.example.stayd.domain.cafe.model.DayOfWeek;
 
 @Getter
 @Builder
